backend/requirements.txt
pymediainfo
imageio
Pillow==10.1.0
uiautomation>=2
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
requests==2.31.0
PyMySQL==1.1.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
whitenoise==6.6.0
setuptools
mysqlclient
sqlalchemy
